[Backend] feat: Test Database Migrations
